requirements.txt
websockets>=12.0
pyperclip>=1.8.2
zeroconf>=0.131.0
python-dotenv>=1.0.1
pytest>=7.0.0
pytest-cov>=4.0.0
keyboard>=0.13.5